+++ src/lib.rs
@@ -1,3 +1,3 @@
-pub mod array2d;
+pub mod fixedvec2d;
 pub mod square;
 pub use square::SquareGraph;
